Fix cells not updating simultaneously
Previously, each cell was updated in a left to right fashion from the
top-left corner to the bottom-right corner of the board. This is wrong,
as the update should be done simultaneously for all the cells. Updating
the cell one by one resulted in the wrong number of neighbors for cells
that are updated later.