Implement more advanced UI with curses
Previously, UI was just printing the board onto console. This is ugly
because
- The symbols used to represent alive or dead cells are just ASCII
characters.
- The board in the previous time step is not being erased, and the board
in the current time step is just being appended directly below. This
results in difficult visualization.

The new UI
- Uses Unicode characters to represent the cells, which is more
aesthetically pleasing.
- The board in the previous time step is erased before printing out the
board in the current time step.
- Follows the MVC [1] pattern, which is much better for reducing
coupling [2]. [3] is a rather good example of MVC.

However, the new UI has some drawbacks.
- An error would be thrown if the console cannot accomodate the size of
the board.

[1]: https://en.wikipedia.org/wiki/Model%E2%80%93view%E2%80%93controller
[2]: https://en.wikipedia.org/wiki/Coupling_(computer_programming)
[3]: https://www.giacomodebidda.com/mvc-pattern-in-python-introduction-and-basicmodel/