Change symbols and remove flicker
Seems like having a problem displaying the following unicode characters
on other terminals:
- BLACK LARGE SQUARE: 2B1B
- WHITE LARGE SQUARE: 2B1C
So replacing these characters.

Also the screen kept flickering, especially for low time delays. Turns
out that this is caused by window.clear() [1], which causes the whole
window to be repainted when the window.refresh() [2] is called. A better
way is to use window.erase() [3], which doesn't cause the whole window
to be repainted. This seems to solve the window flicker.

[1]: https://docs.python.org/3.6/library/curses.html#curses.window.clear
[2]: https://docs.python.org/3.6/library/curses.html#curses.window.refresh
[3]: https://docs.python.org/3.6/library/curses.html#curses.window.erase